+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Regression in local field projection leading to incorrect results for `far_field_approx=True`.
+- Bug when differentiating with respect to `Cylinder.center`.
+
 
 ## [2.7.6] - 2024-10-30